pyproject.toml
[project]
name = "apc-module"
version = "0.1.0"
description = ""
authors = [
    {name = "bict-scpark",email = "scpark@bict.co.kr"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "dependency-injector (>=4.46.0,<5.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "ray (>=2.43.0,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "aiokafka[lz4] (>=0.12.0,<0.13.0)",
    "python-socketio[asgi] (>=5.12.1,<6.0.0)",
    "lz4 (>=4.4.3,<5.0.0)",
    "aiohttp (>=3.11.14,<4.0.0)"
]

[tool.poetry]
packages = [{include = "apc_module", from = "src"}]

[tool.poe.tasks]
format = "black ."
lint = "ruff check --fix ."
dev = "uvicorn src.apc_module.main:app"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
ruff = "^0.9.8"
pre-commit = "^4.1.0"
poethepoet = "^0.32.2"
pyright = "^1.1.396"
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"

[tool.ruff.lint]
select = ["E", "F", "N"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
